Handle host with unset ansible_host
We are having all machines in DNS and want to be able to change IP addresses in DNS. So we do not 
use ansible_host in our host_vars/machine.yml

As os_keystone is the first Ansible role we use. We will make similar changes to other roles later 
on.

Change-Id: Ic9f43cc3f6b62b5098e85afcf55f008c022517f6