Fix keystone_httpd_mpm_max_requests default value
At the moment keystone_httpd_mpm_server_limit appears as string,
so instead of multiplication of integers we repeat value of
`keystone_httpd_mpm_server_limit` `keystone_httpd_mpm_thread_child` times
which is not what we want to do.
So we apply int filter to ensure we do math operation on integers.

Change-Id: Ib3258eb018f758edb9a6a9424a7be8266c7e9fd6