Ensure correct order for credential rotate/migrate
Accroding to the note in keystone-manage code the proper order to
execute credetial rotation is to perform rotation first and migrate to
the new private key afterwars.
Our current code was doing vice versa for now. While it should not lead
to any issues as our autorotate script would fix that later on, let's
still improve task ordering and try to catch credential rotation issues
in ansible code as well, not only in autorotate cron job.

[1] https://opendev.org/openstack/keystone/src/commit/f45921840c17e3b967a136de594f0c7eade82488/keystone/cmd/cli.py#L803-L830

Related-Bug: #2074196
Change-Id: I231cd6ddbfe837ed590c16c806023075102cc23d